+++ src/transformation/visitors/return.ts
@@ -1,12 +1,51 @@
 import * as ts from "typescript";
 import * as lua from "../../LuaAST";
-import { FunctionVisitor } from "../context";
+import { FunctionVisitor, TransformationContext } from "../context";
 import { isInTupleReturnFunction, isTupleReturnCall } from "../utils/annotations";
 import { validateAssignment } from "../utils/assignment-validation";
 import { createUnpackCall, wrapInTable } from "../utils/lua-ast";
 import { ScopeType, walkScopesUp } from "../utils/scope";
 import { isArrayType } from "../utils/typescript";
 import { transformMultiReturnStatement } from "./language-extensions/multi";
+
+function transformExpressionsInReturn(
+    context: TransformationContext,
+    node: ts.Expression,
+    insideTryCatch: boolean
+): lua.Expression[] {
+    if (!isInTupleReturnFunction(context, node)) {
+        return [context.transformExpression(node)];
+    }
+
+    let results: lua.Expression[];
+    const expressionType = context.checker.getTypeAtLocation(node);
+
+    // Parent function is a TupleReturn function
+    if (ts.isArrayLiteralExpression(node)) {
+        // If return expression is an array literal, leave out brackets.
+        results = node.elements.map(e => context.transformExpression(e));
+    } else if (!isTupleReturnCall(context, node) && isArrayType(context, expressionType)) {
+        // If return expression is an array-type and not another TupleReturn call, unpack it
+        results = [createUnpackCall(context, context.transformExpression(node), node)];
+    } else {
+        results = [context.transformExpression(node)];
+    }
+
+    // Wrap tupleReturn results when returning inside try/catch
+    if (insideTryCatch) {
+        results = [wrapInTable(...results)];
+    }
+
+    return results;
+}
+
+export function transformExpressionBodyToReturnStatement(
+    context: TransformationContext,
+    node: ts.Expression
+): lua.Statement {
+    const expressions = transformExpressionsInReturn(context, node, false);
+    return lua.createReturnStatement(expressions, node);
+}
 
 export const transformReturnStatement: FunctionVisitor<ts.ReturnStatement> = (statement, context) => {
     // Bubble up explicit return flag and check if we're inside a try/catch block
@@ -35,27 +74,7 @@ export const transformReturnStatement: FunctionVisitor<ts.ReturnStatement> = (st
             validateAssignment(context, statement, expressionType, returnType);
         }
 
-        if (isInTupleReturnFunction(context, statement)) {
-            // Parent function is a TupleReturn function
-            if (ts.isArrayLiteralExpression(statement.expression)) {
-                // If return expression is an array literal, leave out brackets.
-                results = statement.expression.elements.map(e => context.transformExpression(e));
-            } else if (!isTupleReturnCall(context, statement.expression) && isArrayType(context, expressionType)) {
-                // If return expression is an array-type and not another TupleReturn call, unpack it
-                results = [
-                    createUnpackCall(context, context.transformExpression(statement.expression), statement.expression),
-                ];
-            } else {
-                results = [context.transformExpression(statement.expression)];
-            }
-
-            // Wrap tupleReturn results when returning inside try/catch
-            if (insideTryCatch) {
-                results = [wrapInTable(...results)];
-            }
-        } else {
-            results = [context.transformExpression(statement.expression)];
-        }
+        results = transformExpressionsInReturn(context, statement.expression, insideTryCatch);
     } else {
         // Empty return
         results = [];