+++ test/unit/classes/decorators.spec.ts
@@ -126,3 +126,61 @@ test("Exported class decorator", () => {
         .setReturnExport("Foo", "bar")
         .expectToMatchJsResult();
 });
+
+test.each([
+    ["@decorator method() {}"],
+    ["@decorator property;"],
+    ["@decorator propertyWithInitializer = () => {};"],
+    ["@decorator ['evaluated property'];"],
+    ["@decorator get getter() { return 5 }"],
+    ["@decorator set setter(value) {}"],
+    ["@decorator static method() {}"],
+    ["@decorator static property;"],
+    ["@decorator static propertyWithInitializer = () => {}"],
+    ["@decorator static get getter() { return 5 }"],
+    ["@decorator static set setter(value) {}"],
+    ["@decorator static ['evaluated property'];"],
+    ["method(@decorator a) {}"],
+    ["static method(@decorator a) {}"],
+])("Decorate class member (%p)", classMember => {
+    util.testFunction`
+        let decoratorParameters: any;
+
+        const decorator = (target, key, index?) => {
+            const targetKind = target === Foo ? "Foo" : target === Foo.prototype ? "Foo.prototype" : "unknown";
+            decoratorParameters = { targetKind, key, index: typeof index };
+        };
+
+        class Foo {
+            ${classMember}
+        }
+
+        return decoratorParameters;
+    `.expectToMatchJsResult();
+});
+
+describe("Decorators /w descriptors", () => {
+    test.each([
+        ["return { writable: true }", "return { configurable: true }"],
+        ["desc.writable = true", "return { configurable: true }"],
+    ])("Combine decorators (%p + %p)", (decorateA, decorateB) => {
+        util.testFunction`
+            const A = (target, key, desc): any => { ${decorateA} };
+            const B = (target, key, desc): any => { ${decorateB} };
+            class Foo { @A @B static method() {} }
+            const { value, ...rest } = Object.getOwnPropertyDescriptor(Foo, "method");
+            return rest;
+        `.expectToMatchJsResult();
+    });
+
+    test.each(["return { value: true }", "desc.value = true"])(
+        "Use decorator to override method value",
+        overrideStatement => {
+            util.testFunction`
+                const decorator = (target, key, desc): any => { ${overrideStatement} };
+                class Foo { @decorator static method() {} }
+                return Foo.method;
+            `.expectToMatchJsResult();
+        }
+    );
+});